adding types and adding the babel plugin-module-resolver
