+++ index.ts
@@ -2,13 +2,17 @@ import axios, { AxiosRequestConfig, Method } from "axios";
 import { Route } from "@custom-types/Route";
 import { KeyValuePair } from "@custom-types/KeyValuePair";
 import { GenericObject } from "@custom-types/GenericObject";
-import { BaseItem } from "@custom-types/BaseItem";
 import { ApplyType } from "@custom-types/ApplyType";
-import { CouponCode } from "@custom-types/CouponCode";
-import { GiftCardCode } from "@custom-types/GiftCardCode";
-import { CreatorCode } from "@custom-types/CreatorCode";
 import { Data } from "@custom-types/Data";
 import { Message } from "@custom-types/Message";
+import { Category } from "@custom-types/Category";
+import { ApplyTypeToInterface } from "@custom-types/ApplyTypeToInterface";
+import { Basket } from "@custom-types/Basket/Basket";
+import { AuthUrl } from "@custom-types/AuthUrl";
+import { PackageType } from "@custom-types/Package/PackageType";
+import { Webstore } from "@custom-types/Webstore";
+import { Page } from "@custom-types/Page";
+import { Package } from "@custom-types/Package/Package";
 
 /**
  * @constant baseUrl
@@ -72,295 +76,6 @@ export async function Request<T, Body>(
     const response = await axios.request<T>(config);
 
     return response.data;
-}
-
-/**
- * @type {Category}
- * @description The category object returned from the Tebex Headless API
- *
- * @param {number} id The ID of the category
- * @param {string} name The name of the category
- * @param {string} description The description of the category
- * @param {Category | null} parent The parent category of the category
- * @param {number} order The order of the category
- * @param {Package[]} packages The packages in the category
- * @param {"grid" | "list" | string} display_type The display type of the category
- * @param {string | null} slug The slug of the category
- */
-export type Category = BaseItem & {
-    description: string;
-    parent: Category | null;
-    order: number;
-    packages: Package[];
-    display_type: "grid" | "list";
-    slug: string | null;
-};
-
-/**
- * @type {ApplyTypeToInterface}
- * @description The type of the apply request
- *
- * @param {string} coupons The coupons type
- * @param {string} giftcards The giftcards type
- * @param {string} creator-codes The creator codes type
- *
- * @returns {CouponCode | GiftCardCode | CreatorCode}
- */
-export type ApplyTypeToInterface<T extends ApplyType> = T extends "coupons"
-    ? CouponCode
-    : T extends "giftcards"
-    ? GiftCardCode
-    : T extends "creator-codes"
-    ? CreatorCode
-    : never;
-
-/**
- * @type {PackageType}
- * @description The type of the package
- *
- * @param {string} subscription The subscription type
- * @param {string} single The single type
- */
-export type PackageType = "subscription" | "single" | "both";
-
-/**
- * @type {Package}
- * @description The package object returned from the Tebex Headless API
- *
- * @param {number} id The ID of the package
- * @param {string} name The name of the package
- * @param {string} description The description of the package
- * @param {PackageType} type The type of the package
- * @param {boolean} disable_gifting Whether gifting is disabled for the package
- * @param {boolean} disable_quantity Whether quantity is disabled for the package
- * @param {string | null} expiration_date The expiration date of the package
- * @param {BaseItem} category The category of the package
- * @param {number} base_price The base price of the package
- * @param {number} sales_tax The sales tax of the package
- * @param {number} total_price The total price of the package
- * @param {number} discount The discount of the package
- * @param {string | null} image The image of the package
- * @param {string} created_at The date the package was created
- * @param {string} updated_at The date the package was updated
- * @param {number} order The order this package should be sorted in
- */
-export type Package = BaseItem & {
-    description: string;
-    type: PackageType;
-    disable_gifting: boolean;
-    disable_quantity: boolean;
-    expiration_date: string | null;
-    currency: string;
-    category: BaseItem;
-    base_price: number;
-    sales_tax: number;
-    total_price: number;
-    discount: number;
-    image: string | null;
-    created_at: string;
-    updated_at: string;
-    order: number;
-};
-
-/**
- * @type {InBasket}
- * @description The in_basket object inside a basket package object
- *
- * @param {number} quantity The quantity of the package in the basket
- * @param {number} price The price of the package in the basket
- * @param {string | null} gift_username_id The ID of the user the package is gifted to
- * @param {string | null} gift_username The username of the user the package is gifted to
- */
-export type InBasket = {
-    quantity: number;
-    price: number;
-    gift_username_id: string | null;
-    gift_username: string | null;
-}
-
-/**
- * @type {BasketPackage}
- * @description The basket package object returned from the Tebex Headless API
- *
- * @param {number} id The ID of the package
- * @param {string} name The name of the package
- * @param {string} description The description of the package
- * @param {InBasket} in_basket The in_basket object inside the basket package object
- * @param {string | null} image The image of the package
- */
-export type BasketPackage = BaseItem & {
-    description: string;
-    in_basket: InBasket;
-    image: string | null;
-};
-
-/**
- * @type {Code}
- * @description The code object inside the basket coupons object
- *
- * @param {string} code The code of the object
- */
-export type Code = {
-    code: string;
-}
-
-/**
- * @type {Links}
- * @description The links object inside the basket object
- *
- * @param {string} checkout The checkout link of the basket
- */
-export type Links = {
-    checkout: string;
-    [key: string]: string;
-}
-
-/**
- * @type {Basket}
- * @description The basket object returned from the Tebex Headless API
- *
- * @param {string} ident The identifier of the basket
- * @param {boolean} complete Whether the basket is complete
- * @param {number} id The ID of the basket
- * @param {string} country The country of the basket
- * @param {string} ip The IP address of the user
- * @param {string | null} username_id The ID of the user
- * @param {string | null} username The username of the user
- * @param {string} cancel_url The cancel url of the basket
- * @param {string} complete_url The complete url of the basket
- * @param {boolean} complete_auto_redirect Whether the basket should automatically redirect to the complete url
- * @param {number} base_price The base price of the basket
- * @param {number} sales_tax The sales tax of the basket
- * @param {number} total_price The total price of the basket
- * @param {string} email The email of the basket
- * @param {string} currency The currency of the basket
- * @param {BasketPackage[]} packages The packages in the basket
- * @param {Code[]} coupons The coupons in the basket
- * @param {GiftCardCode[]} giftcards The giftcards in the basket
- * @param {string} creator_code The creator code of the basket
- * @param {Links} links The links of the basket
- * @param {KeyValuePair<string, any>} custom The custom object of the basket
- */
-export type Basket = {
-    ident: string;
-    complete: boolean;
-    id: number;
-    country: string;
-    ip: string;
-    username_id: string | null;
-    username: string | null;
-    cancel_url: string;
-    complete_url: string;
-    complete_auto_redirect: boolean;
-    base_price: number;
-    sales_tax: number;
-    total_price: number;
-    email: string;
-    currency: string;
-    packages: BasketPackage[];
-    coupons: Code[];
-    giftcards: GiftCardCode[];
-    creator_code: string;
-    links: Links;
-    custom: KeyValuePair<string, any>;
-}
-
-/**
- * @type {Urls}
- * @description The url object for the complete and cancel urls
- *
- * @param {string} complete_url The complete url
- * @param {string} cancel_url The cancel url
- */
-export type Urls = {
-    complete_url: string;
-    cancel_url: string;
-    custom?: KeyValuePair<string, any>;
-    complete_auto_redirect?: boolean;
-};
-
-/**
- * @type {AuthUrl}
- * @description The auth url object returned from the Tebex Headless API
- *
- * @param {string} name The name of the auth url
- * @param {string} url The url of the auth url
- */
-export type AuthUrl = {
-    name: string;
-    url: string;
-}
-
-/**
- * @type {PackageBody}
- * @description The package object for the body of the request
- *
- * @param {number} package_id The ID of the package
- * @param {number} quantity The quantity of the package
- * @param {PackageType} type The type of the package
- */
-export type PackageBody = {
-    package_id: number;
-    quantity: number;
-    type: PackageType;
-}
-
-/**
- * @type {Webstore}
- * @description The webstore object returned from the Tebex Headless API
- *
- * @param {number} id The ID of the webstore
- * @param {string} description The description of the webstore
- * @param {string} name The name of the webstore
- * @param {string} webstore_url The webstore url of the webstore
- * @param {string} currency The currency of the webstore
- * @param {string} lang The language of the webstore
- * @param {string} logo The logo of the webstore
- * @param {string} platform_type The platform type of the webstore
- * @param {number} platform_type_id The platform type ID of the webstore
- * @param {string} created_at The date the webstore was created
- */
-export type Webstore = {
-    id: number;
-    description: string;
-    name: string;
-    webstore_url: string;
-    currency: string;
-    lang: string;
-    logo: string;
-    platform_type: string;
-    platform_type_id: number;
-    created_at: string;
-}
-
-/**
- * @type {Page}
- * @description The page object returned from the Tebex Headless API
- * 
- * @param {number} id The ID of the page
- * @param {string} created_at The date the page was created
- * @param {string} updated_at The date the page was updated
- * @param {number} account_id The ID of the account
- * @param {string} title The title of the page
- * @param {string} slug The slug of the page
- * @param {boolean} private Whether the page is private
- * @param {boolean} hidden Whether the page is hidden
- * @param {boolean} disabled Whether the page is disabled
- * @param {boolean} sequence Whether the page is in a sequence
- * @param {string} content The content of the page
-*/
-export type Page = {
-    id: number;
-    created_at: string;
-    updated_at: string;
-    account_id: number;
-    title: string;
-    slug: string;
-    private: boolean;
-    hidden: boolean;
-    disabled: boolean;
-    sequence: boolean;
-    content: string;
 }
 
 export class TebexHeadless {
