+++ package.json
@@ -28,8 +28,8 @@
     },
     "homepage": "https://github.com/Tynopia/tebex_headless#readme",
     "devDependencies": {
-        "@babel/preset-typescript": "^7.23.3",
         "@babel/preset-env": "^7.23.6",
+        "@babel/preset-typescript": "^7.23.3",
         "@swc/core": "^1.3.99",
         "@types/jest": "^29.5.11",
         "dotenv": "^16.3.1",
@@ -40,7 +40,8 @@
         "typescript": "^5.3.2"
     },
     "dependencies": {
-        "axios": "^1.6.2"
+        "axios": "^1.6.2",
+        "babel-plugin-module-resolver": "^5.0.2"
     },
     "publishConfig": {
         "registry": "https://registry.npmjs.org/",
