+++ src/@types/AuthUrl.ts
@@ -1,0 +1,12 @@
+
+/**
+ * @type {AuthUrl}
+ * @description The auth url object returned from the Tebex Headless API
+ *
+ * @param {string} name The name of the auth url
+ * @param {string} url The url of the auth url
+ */
+export type AuthUrl = {
+    name: string;
+    url: string;
+}
