+++ src/@types/Code/Code.ts
@@ -1,0 +1,9 @@
+/**
+ * @type {Code}
+ * @description The code object inside the basket coupons object
+ *
+ * @param {string} code The code of the object
+ */
+export type Code = {
+    code: string;
+}
