+++ src/@types/Package/PackageType.ts
@@ -1,0 +1,8 @@
+/**
+ * @type {PackageType}
+ * @description The type of the package
+ *
+ * @param {string} subscription The subscription type
+ * @param {string} single The single type
+ */
+export type PackageType = "subscription" | "single" | "both";
